+++ lib/shopify-cli/http_request.rb
@@ -2,14 +2,26 @@ require 'net/http'
 
 module ShopifyCli
   class HttpRequest
-    def self.call(uri, body, variables, headers)
-      http = ::Net::HTTP.new(uri.host, uri.port)
-      http.use_ssl = true
-      req = ::Net::HTTP::Post.new(uri.request_uri)
-      req.body = JSON.dump(query: body.tr("\n", ""), variables: variables)
-      req['Content-Type'] = 'application/json'
-      headers.each { |header, value| req[header] = value }
-      http.request(req)
+    class << self
+      def post(uri, body, headers)
+        req = ::Net::HTTP::Post.new(uri.request_uri)
+        request(uri, body, headers, req)
+      end
+
+      def get(uri, body, headers)
+        req = ::Net::HTTP::Get.new(uri.request_uri)
+        request(uri, body, headers, req)
+      end
+
+      def request(uri, body, headers, req)
+        http = ::Net::HTTP.new(uri.host, uri.port)
+        http.use_ssl = true
+
+        req.body = body unless body.nil?
+        req['Content-Type'] = 'application/json'
+        headers.each { |header, value| req[header] = value }
+        http.request(req)
+      end
     end
   end
 end